+++ python/ray/serve/_private/common.py
@@ -395,5 +395,7 @@ class MultiplexedReplicaInfo:
 
 @dataclass
 class StreamingHTTPRequest:
+    """Sent from the HTTP proxy to replicas on the streaming codepath."""
+
     pickled_asgi_scope: bytes
     http_proxy_handle: ActorHandle
