+++ python/ray/serve/_private/utils.py
@@ -11,6 +11,7 @@ from enum import Enum
 from functools import wraps
 from typing import (
     Any,
+    AsyncGenerator,
     Callable,
     Dict,
     Iterable,
@@ -723,3 +724,24 @@ def calculate_remaining_timeout(
 
     time_since_start_s = curr_time_s - start_time_s
     return max(0, timeout_s - time_since_start_s)
+
+
+def wrap_generator_function_in_async_if_needed(
+    f: Callable,
+) -> Callable[[Any], AsyncGenerator]:
+    """Given a callable, make sure it returns an async generator.
+
+    If the callable is not a generator at all, raise a `TypeError`.
+    """
+    if inspect.isasyncgenfunction(f):
+        return f
+    elif inspect.isgeneratorfunction(f):
+
+        @wraps(f)
+        async def async_gen_wrapper(*args, **kwargs):
+            for result in f(*args, **kwargs):
+                yield result
+
+        return async_gen_wrapper
+    else:
+        raise TypeError(f"Method '{f.__name__}' is not a generator.")
